added to package to tinacms
