+++ packages/@tinacms/schema-tools/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "@tinacms/schema-tools",
+  "version": "0.0.1",
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "license": "Apache-2.0",
+  "buildConfig": {
+    "entryPoints": [
+      {
+        "name": "src/index.ts",
+        "target": "node"
+      }
+    ]
+  },
+  "devDependencies": {
+    "@tinacms/scripts": "workspace:*",
+    "@types/yup": "^0.29.10",
+    "jest": "^27.0.6",
+    "typescript": "^4.3.5"
+  },
+  "peerDependencies": {
+    "yup": "^0.32.0"
+  },
+  "scripts": {
+    "build": "echo \"Run `yarn build` from the root of the repository instead\"",
+    "test": "jest --passWithNoTests",
+    "types": "yarn tsc",
+    "test-watch": "jest  --passWithNoTests --watch",
+    "generate:schema": "yarn node scripts/generateSchema.js"
+  },
+  "publishConfig": {
+    "registry": "https://registry.npmjs.org"
+  },
+  "repository": {
+    "url": "https://github.com/tinacms/tinacms.git",
+    "directory": "packages/@tinacms/cli"
+  }
+}
