+++ packages/@tinacms/schema-tools/src/schema/index.ts
@@ -1,0 +1,4 @@
+export * from './addNamespaceToSchema'
+export * from './TinaSchema'
+export * from './resolveField'
+export * from './resolveForm'
