+++ packages/@tinacms/schema-tools/src/util/index.ts
@@ -1,0 +1,4 @@
+export * from './validate'
+export * from './lastItem'
+export * from './namer'
+export * from './sequential'
