+++ packages/@tinacms/schema-tools/src/util/sequential.ts
@@ -1,0 +1,38 @@
+/**
+ * Iterate through an array of promises sequentially, ensuring the order
+ * is preserved.
+ *
+ * ```js
+ * await sequential(templates, async (template) => {
+ *   await doSomething(template)
+ * })
+ * ```
+ */
+export const sequential = async <A, B>(
+  items: A[] | undefined,
+  callback: (args: A, idx: number) => Promise<B>
+) => {
+  const accum: B[] = []
+  if (!items) {
+    return []
+  }
+
+  const reducePromises = async (previous: Promise<B>, endpoint: A) => {
+    const prev = await previous
+    // initial value will be undefined
+    if (prev) {
+      accum.push(prev)
+    }
+
+    return callback(endpoint, accum.length)
+  }
+
+  // @ts-ignore FIXME: this can be properly typed
+  const result = await items.reduce(reducePromises, Promise.resolve())
+  if (result) {
+    // @ts-ignore FIXME: this can be properly typed
+    accum.push(result)
+  }
+
+  return accum
+}
