+++ packages/tinacms/src/hooks/use-graphql-forms.tsx
@@ -13,7 +13,6 @@ limitations under the License.
 
 import React from 'react'
 import gql from 'graphql-tag'
-import { print } from 'graphql'
 import { getIn, setIn } from 'final-form'
 import {
   useCMS,
@@ -27,7 +26,7 @@ import { assertShape, safeAssertShape } from '../utils'
 import type { FormOptions, TinaCMS } from '@tinacms/toolkit'
 import { BiLinkExternal } from 'react-icons/bi'
 import { useFormify } from './formify'
-import { TinaSchema, resolveForm } from '../schema'
+import { TinaSchema, resolveForm } from '@tinacms/schema-tools'
 
 export function useGraphqlFormsUnstable<T extends object>({
   variables,
@@ -236,30 +235,30 @@ export function useGraphqlForms<T extends object>({
           )
 
           // Uncomment this to test this work.
-          // const enrichedSchema: TinaSchema = cms.api.tina.schema
-          // const collection = enrichedSchema.getCollection(
-          //   result._internalSys.collection.name
-          // )
-          // const template = await enrichedSchema.getTemplateForData({
-          //   collection,
-          //   data: result.values,
-          // })
+          const enrichedSchema: TinaSchema = cms.api.tina.schema
+          const collection = enrichedSchema.getCollection(
+            result._internalSys.collection.name
+          )
+          const template = await enrichedSchema.getTemplateForData({
+            collection,
+            data: result.values,
+          })
 
-          // const formInfo = await resolveForm({
-          //   collection,
-          //   basename: collection.name,
-          //   schema: enrichedSchema,
-          //   template,
-          // })
+          const formInfo = await resolveForm({
+            collection,
+            basename: collection.name,
+            schema: enrichedSchema,
+            template,
+          })
 
           const formConfig = {
             id: queryName,
             initialValues: result.values,
-            label: result.form.label,
-            fields: result.form.fields,
+            // label: result.form.label,
+            // fields: result.form.fields,
             // Uncomment this to test this work.
-            // label: formInfo.label,
-            // fields: formInfo.fields,
+            label: formInfo.label,
+            fields: formInfo.fields,
 
             reset: () => {
               setPendingReset(queryName)
